Move ChangePassword component to separate file
