+++ client/src/Components/Settings/Settings.js
@@ -1,15 +1,11 @@
-import React from 'react';
-import axios from 'axios';
+import React, { Component } from 'react';
 
 import ChangeEmail from './ChangeEmail';
+import ChangePassword from './ChangePassword';
 
-const BACKEND_URL = process.env.REACT_APP_BACKEND_URL;
 
-class Settings extends React.Component {
+class Settings extends Component {
   state = {
-    currentPassword: '',
-    newPassword: '',
-    confirmNewPassword: '',
     autoScratch: false
   };
 
@@ -24,80 +20,11 @@ class Settings extends React.Component {
     });
   };
 
-  handleSubmitPassword = async e => {
-    e.preventDefault();
-    console.log(this.state);
-
-    const { currentPassword, newPassword, confirmNewPassword } = this.state;
-
-    // Error handling
-    if (!currentPassword) return console.error('Enter your current password!');
-    if (newPassword !== confirmNewPassword)
-      return console.error('Passwords do not match!');
-    if (newPassword.length < 6)
-      return console.error('Password needs to be at least 6 characters!');
-
-    const url = `${BACKEND_URL}/api/change_password`;
-    // TODO: pull username from global state and send in body
-    const body = { password: currentPassword, new_password: newPassword };
-    const options = {
-      headers: { Authorization: `Bearer ${localStorage.getItem('jwt_token')}` }
-    };
-
-    try {
-      const updatedUser = await axios.post(url, body, options);
-      console.log('Password updated successfully', updatedUser);
-    } catch (err) {
-      console.error('Error changing your password', err);
-    }
-  };
-
-  validateForm() {
-    if (!this.state['email'] || this.state['email'].length === 0) {
-      console.error('Form not validated');
-    }
-  }
-
   render() {
     return (
       <div className="Settings">
         <ChangeEmail />
-        {/*Change password form*/}
-        <form
-          onSubmit={this.handleSubmitPassword}
-          className="Settings__ChangePassword"
-        >
-          <label htmlFor="currentPassword">
-            Current Password
-            <input
-              type="text"
-              name="currentPassword"
-              placeholder="Current password"
-              onChange={e => this.handleChange(e)}
-            />
-          </label>
-
-          <label htmlFor="newPassword">
-            New Password
-            <input
-              type="text"
-              name="newPassword"
-              placeholder="New password"
-              onChange={e => this.handleChange(e)}
-            />
-          </label>
-
-          <label htmlFor="confirmNewPassword">
-            Confirm New password
-            <input
-              type="text"
-              name="confirmNewPassword"
-              placeholder="Confirm new password"
-              onChange={e => this.handleChange(e)}
-            />
-          </label>
-          <input type="submit" placeholder="Submit" />
-        </form>
+        <ChangePassword />
 
         {/*Preferences*/}
         <div>
